Add support for implicit conversions to C++ types
See issue #259.

A new py::implicitly_cpp_convertible<From,To>() tells pybind() that From
can be implicitly converted to To (in C++), and thus pybind11-registered
From types can be used in methods that accept (C++) To instances.

In essence, this lets you make use of C++ implicit converters (typically
via a 'From::operator To()' or a non-explicit 'To::To(const From&)'
constructor) with pybind11 instances without needing to create an
explicit pybind11 interface for the conversion.

As a simple example, consider the two classes:

    class A {
        ...
        void method(uint64_t val) { ... }
    }
    class B {
        ...
        operator uint64_t() { ... } // returns unique identifier
    }

In C++, you can call `a.method(22)` or `a.method(b)`.  Telling
pybind11 about the conversion allows you to do the same in python.

Without the implicit conversion, you would have to either overload
A.method to accept both uint64_t and B instances, or would have to add
a method to B's interface that returns the id, then pass this method
result to A's method.