Merge py::implicitly_(cpp_)convertible interfaces
It's a cleaner design to just have one py::implicitly_convertible<A,B>()
interface.  It works as before (doing python type
conversion/construction) when B is a pybind-registered type, with the
C++ implicit conversion taking over when B is not pybind-registered.

The previous C++-conversion interface allowed use of C++-level implicit
conversion even of pybind11-registered types (for example, if implicit
conversion exist but isn't exposed to the python interface).  I'm not
sure that this was all that useful, but if so, it should still be
possible by putting the implicitly_convertible<A,B>() call after A has
been registered but before B has been registered.  (Once B is
registered, this declares a python-constructor conversion operation).