Move load-or-convert logic from tuple to type_casters
This is an experiment to see how it affects the library size versus the
previous way of keeping most of the code in the tuple type caster.

It doesn't seem to be a big improvement--it reduces the size (excluding
the new example18 test) on my debian g++-6 system to 13.4% from 14.3%,
at the cost of some increased maintenance complexity.