Much reduced so usage for implicit conversion
The biggest shared object use of the previous version was coming from
needing to have argument type information available to perform the
implicit conversion.  That was previously done by requiring a templated
implicit conversion function, but that meant a new function
instantiation and function call for every possibly argument type.
Similar attempts to encoding the argument type_index values in an array
could only slightly reduce the cost (it still required considerable
compile-time storage to allocate the array/vector of types, and had to
be allocated independently for each and every function signature.

This commit takes a more clever approach of encoding argument types
through the name() mechanism: instead of only storing the type_info
reference for un-named types (i.e. types with a '%' value in their
name), this now also encodes the type_info for other types where
implicit conversion should be supported, and then extracts the type
information in cpp_function::initialize_generic(), i.e. outside of
templated code.

This also puts the current set of implicit instances into a stack held
in the pybind11 internals, managing the stack again in non-templated
code.  (The stack is necessary because implicit conversion can easily
recurse, and we need one set of implicit conversion instances per
active call).  The only additions to templated code are a call to
(non-templated) detail::try_implicit_conversion() if regular argument
loading fails (to attempt to satisfy any failed arguments through
implicit C++ conversion), and a non-templated call inside the tuple
type_caster call() method to get the implicit instances at the top of
the implicit instance stack.

My testing shows this adds a constant 15K to the shared object size
(tested by looking at the different in .so size for a source file with
just a single pybind11 function definition under master and this
branch), plus 64 bytes per unique function signature.  For the examples
(with example20 disabled) that works out to a (variable) cost of 2.9%.
Under c++11, the overhead is a little higher: about 16K constant plus
about 4.3%.