Cache return types of called functions
Each time a function is called, we currently do a type lookup in
registered_types_cpp to figure out the python type info associated with
the return value.  This lookup isn't free, particularly for
often-called functions, but will yield and identical result each time
the function is called.

This commit adds an (optional) cache parameter to the lookup for generic
types (i.e. those inheriting from type_caster_generic) so that these
lookups can be cached.