Add support for passing argument details to type_caster load()
This adds support for passing py::arg_v objects to type_casters that
define a `load(handle, bool, py::arg_v)`.  This allows type_casters to
determine more about their arguments, and allows flags on arguments to
be able to signal that different arguments should be treated in
different ways (e.g. allowing a sort of "argument policy" counterpart to
return_value_policy).

To show this working (and test it), I added a "flag_nocopy" that lets
you do things like:

    m.def("f", &f, "a"_a.nocopy() = "default"); // Named, default, nocopy
    m.def("g", &g, py::arg().nocopy()); // Unnammed, no default, nocopy

I intend to use this nocopy flag with eigen to allow users to signal
that an argument shouldn't be copied to make it fit in a Ref<> type
(i.e. raise an exception instead).

This also changes the the py::arg (and thus py::arg_v) classes to
allowing being declared without a name, so that you can specify argument
policies without making an argument a keyword argument.