Split up single char and c-style string conversion
This moves the `char *` caster code into a new `type_caster<char*>` and
repurposes the `type_caster<char>` for doing only single-character
casting (plus the same for the other standard char types).

For common types, treating `T` and `T*` the same makes sense, but `char`
and `char*` really don't work being treated the same at all: a `char*`
is fundamentally different in use compared to a char.

This new caster is done by adding a conditional to `make_caster` that
specifically adds the pointer when casting a `char *` (or similar)
instead of throwing away the pointer (that is, it strips away everything
down to the intrinsic_t, but then re-adds the pointer for these C-style
string types).

In the end, this leaves us now with three sets of related casters
instead of two:

- `std::string`/`std::wstring`/`std::uNNstring`
- `char*`/`wchar_t*`/`charNN_t*`
- `char`/`wchar_t`/`charNN_t`

where the second ones are just wrappers around the first one, but with
extra support for nullptr <-> None conversion.  The third one is all
new: it handles loading and casting single characters, with load
failures for attempting to load a character too large for the associated
type, and works on unicode codepoint inputs, not individual codes.