Clean up pointer type_caster
Small cleanup to the make_caster logic determining whether to make a
`type_caster<T>` or `type_caster<T*>` via partial specialization.
Currently the specialization is only defined for char types, but
something else with a need to do the same in the future could similarly
provide a specialization rather than this being char-specific logic.