Fail to compile with MI via class_ ctor parameters
We can't support this for classes from imported modules (which is the
primary purpose of a ctor argument base class) because we *have* to
have both parent and derived to properly extract a multiple-inheritance
base class pointer from a derived class pointer.

We could support this for actual `class_<Base, ...> instances, but since
in that case the `Base` is already present in the code, it seems more
consistent to simply always require MI to go via template options.

This also adds a new "number_of" template, mostly based on the `count_t`
removed in PR #554, to count the number of satisfied templates; this is
basically just `constexpr_sum`, and uses that except for under MSVC
which doesn't think the constexpr is a constexpr.