travis-ci: build on trusty; switch to python 3.6
This updates the travis-ci non-docker builds to use the trusty image
rather than the default precise build.  This means we don't need to pull
backported versions of the compiler toolchain and cmake (essentially the
backport ppas were giving us the trusty versions in precise), which
should speed up the builds as they no longer have to download packages.

This also updates the gcc 4.8 build to python 3.6 (which is available
in deadsnakes for trusty, but not for precise).

The BAREBONES build now builds with 3.5 as the trusty image has it
installed (precise didn't) and pybind is picking it up over python 2.7
(as it should!).