Fix class name in alias constructors
For classes with an alias, the type shows up in the constructor as
"handle" rather than the actual type, because py::init<> uses a
`handle` first argument in such a case because it needs to make a
run-time decision as to whether to construct a Base or Alias type.

This fixes the signature by replacing the first argument type the scope
(i.e. module) name when the first argument of an instance method is set
to `handle`.  This also ends up fixing the `self` argument of
non-constructor methods taking `self` as a handle, which seems less
important, but still desirable.