Allow binding factory functions as constructors
This allows you to use:

    cls.def(py::init_factory(&factory_function));

where `factory_function` is some pointer, holder, value, or
handle-generating factory function of the type that `cls` binds.

Various compile-time checks are performed to ensure the function is
valid, and various run-time type checks where necessary (i.e. when using
a python-object-returning function, or when a dynamic_cast is needed
for downcasting a pointer).

The feature is optional, and requires including the <pybind11/factory.h>
header.