Reduce instance size; other small optimizations
This reduces the size of the instance by taking the all_type_info_data
out of it; instead just a single `void *` which is either a
detail::type_info or a std::vector* is kept to cache the type info data.

This also simplifies some of the value_and_holder iteration code to
slightly improve performance.