Enhance py::bytes interface to allow extracting raw data
Currently the only way to get data out of a `py::bytes` is to copy it
into a `std::string`.  This commit adds a few alternatives:

- implicit conversion to a `std::string_view` which references the
  Python-held buffer (rather than a copy)
- extracting a `const char *` reference to the raw data via a new
  `data()` method; the size of the referenced data can be obtained via a
  new `size()` method.

This also adds implicit construction from `std::string_view` (though
this still involves a copy) for symmetry.

In addition to testing the above features, this also adds a couple of
missing tests of the already-existing `char *` and `char *, size_t`
constructors of `py::bytes`.

(The `std::string_view` support is only turned on when compiling in
C++17 mode, of course).